🚧 WIP: Started work on simplifying responses - Responses & Sessions are now merged to single models - Responses now track multiple questions' answers in json data object - TODO: Update routes
